+++ forui/example/lib/example.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
 
+import 'package:forui/forui.dart';
+
 class Example extends StatefulWidget {
   const Example({super.key});
 
@@ -14,5 +16,14 @@ class _ExampleState extends State<Example> {
   }
 
   @override
-  Widget build(BuildContext context) => const Placeholder();
+  Widget build(BuildContext context) => Column(
+        children: [
+          FCalendar(
+            controller: FCalendarValueController(),
+            initialType: FCalendarPickerType.yearMonth,
+            start: DateTime.utc(2000),
+            end: DateTime.utc(2030),
+          ),
+        ],
+      );
 }