+++ forui/lib/src/foundation/tappable.dart
@@ -1,11 +1,16 @@
 import 'package:flutter/foundation.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 
 import 'package:meta/meta.dart';
 
+// TODO: Remove redundant comment when flutter fixes its lint issue.
+///
+@internal
+typedef FTappableState = ({bool focused, bool hovered});
+
 @internal
 class FTappable extends StatefulWidget {
-  final bool enabled;
   final String? semanticLabel;
   final bool selected;
   final bool excludeSemantics;
@@ -14,7 +19,7 @@ class FTappable extends StatefulWidget {
   final ValueChanged<bool>? onFocusChange;
   final VoidCallback? onPress;
   final VoidCallback? onLongPress;
-  final ValueWidgetBuilder<bool> builder;
+  final ValueWidgetBuilder<FTappableState> builder;
   final Widget? child;
 
   factory FTappable.animated({
@@ -26,7 +31,7 @@ class FTappable extends StatefulWidget {
     ValueChanged<bool>? onFocusChange,
     VoidCallback? onPress,
     VoidCallback? onLongPress,
-    ValueWidgetBuilder<bool>? builder,
+    ValueWidgetBuilder<FTappableState>? builder,
     Widget? child,
     Key? key,
   }) = _AnimatedTappable;
@@ -40,12 +45,13 @@ class FTappable extends StatefulWidget {
     this.onFocusChange,
     this.onPress,
     this.onLongPress,
-    ValueWidgetBuilder<bool>? builder,
+    ValueWidgetBuilder<FTappableState>? builder,
     this.child,
     super.key,
   })  : assert(builder != null || child != null, 'Either builder or child must be provided.'),
-        builder = builder ?? ((_, __, child) => child!),
-        enabled = onPress != null || onLongPress != null;
+        builder = builder ?? ((_, __, child) => child!);
+
+  bool get enabled => onPress != null || onLongPress != null;
 
   @override
   State<FTappable> createState() => _FTappableState();
@@ -80,33 +86,51 @@ class _FTappableState extends State<FTappable> with SingleTickerProviderStateMix
   bool _hovered = false;
 
   @override
-  Widget build(BuildContext context) => Semantics(
-        enabled: widget.enabled,
-        label: widget.semanticLabel,
-        container: true,
-        button: true,
-        selected: widget.selected,
-        excludeSemantics: widget.excludeSemantics,
-        child: Focus(
-          autofocus: widget.autofocus,
-          focusNode: widget.focusNode,
-          onFocusChange: (focused) {
-            setState(() => _focused = focused);
-            widget.onFocusChange?.call(focused);
-          },
-          child: MouseRegion(
-            cursor: widget.enabled ? SystemMouseCursors.click : MouseCursor.defer,
-            onEnter: (_) => setState(() => _hovered = true),
-            onExit: (_) => setState(() => _hovered = false),
-            child: _child,
-          ),
+  Widget build(BuildContext context) {
+    final tappable = Semantics(
+      enabled: widget.enabled,
+      label: widget.semanticLabel,
+      container: true,
+      button: true,
+      selected: widget.selected,
+      excludeSemantics: widget.excludeSemantics,
+      child: Focus(
+        autofocus: widget.autofocus,
+        focusNode: widget.focusNode,
+        onFocusChange: (focused) {
+          setState(() => _focused = focused);
+          widget.onFocusChange?.call(focused);
+        },
+        child: MouseRegion(
+          cursor: widget.enabled ? SystemMouseCursors.click : MouseCursor.defer,
+          onEnter: (_) => setState(() => _hovered = true),
+          onExit: (_) => setState(() => _hovered = false),
+          child: _child,
         ),
-      );
+      ),
+    );
+
+    if (widget.onPress == null) {
+      return tappable;
+    }
+
+    return Shortcuts(
+      shortcuts: const {
+        SingleActivator(LogicalKeyboardKey.enter): ActivateIntent(),
+      },
+      child: Actions(
+        actions: {
+          ActivateIntent: CallbackAction<ActivateIntent>(onInvoke: (_) => widget.onPress!()),
+        },
+        child: tappable,
+      ),
+    );
+  }
 
   Widget get _child => GestureDetector(
         onTap: widget.onPress,
         onLongPress: widget.onLongPress,
-        child: widget.builder(context, _focused || _hovered, widget.child),
+        child: widget.builder(context, (focused: _focused, hovered: _hovered), widget.child),
       );
 }
 
@@ -156,7 +180,7 @@ class _AnimatedTappableState extends _FTappableState {
                   _controller.forward();
                 },
           onLongPress: widget.onLongPress,
-          child: widget.builder(context, _focused || _hovered, widget.child),
+          child: widget.builder(context, (focused: _focused, hovered: _hovered), widget.child),
         ),
       );