+++ forui/lib/src/widgets/calendar/month/month_picker.dart
@@ -5,7 +5,9 @@ import 'package:intl/intl.dart';
 import 'package:meta/meta.dart';
 import 'package:sugar/sugar.dart';
 
+import 'package:forui/src/widgets/calendar/day/day_picker.dart';
 import 'package:forui/src/widgets/calendar/shared/entry.dart';
+import 'package:forui/src/widgets/calendar/year/year_picker.dart';
 import 'package:forui/src/widgets/calendar/year_month_picker.dart';
 
 // ignore: non_constant_identifier_names
@@ -64,23 +66,27 @@ class _MonthPickerState extends State<MonthPicker> {
   }
 
   @override
-  Widget build(BuildContext context) => GridView(
-        gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
-          crossAxisCount: MonthPicker.columns,
-          childAspectRatio: 1.618,
+  Widget build(BuildContext context) => Padding(
+        padding: const EdgeInsets.only(top: 5.0),
+        child: GridView(
+          gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
+            crossAxisCount: YearPicker.columns,
+            mainAxisExtent: ((DayPicker.tileDimension - 5.0) * DayPicker.maxRows) / YearPicker.rows,
+            mainAxisSpacing: 5.0,
+          ),
+          children: [
+            for (var month = widget.currentYear, i = 0; i < 12; month = month.plus(months: 1), i++)
+              Entry.yearMonth(
+                style: widget.style,
+                date: month,
+                focusNode: _months[i],
+                current: widget.today.truncate(to: DateUnit.months) == month,
+                selectable: widget.start <= month && month <= widget.end,
+                format: (date) => _MMM.format(date.toNative()), // TODO: localize
+                onPress: widget.onPress,
+              ),
+          ],
         ),
-        children: [
-          for (var month = widget.currentYear, i = 0; i < 12; month = month.plus(months: 1), i++)
-            Entry.yearMonth(
-              style: widget.style,
-              date: month,
-              focusNode: _months[i],
-              current: widget.today.truncate(to: DateUnit.months) == month,
-              enabled: widget.start <= month && month <= widget.end,
-              format: (date) => _MMM.format(date.toNative()), // TODO: localize
-              onPress: widget.onPress,
-            ),
-        ],
       );
 
   @override